fix: relocate bench_utils ckd functions
